+++ pom.xml
@@ -54,12 +54,6 @@
             <artifactId>lombok</artifactId>
             <version>1.18.26</version>
         </dependency>
-        <!-- https://mvnrepository.com/artifact/org.liquibase/liquibase-core -->
-        <dependency>
-            <groupId>org.liquibase</groupId>
-            <artifactId>liquibase-core</artifactId>
-            <version>4.18.0</version>
-        </dependency>
     </dependencies>
 
     <build>
