+++ src/main/resources/application.properties
@@ -6,4 +6,3 @@ spring.jpa.show-sql=true
 
 path.to.avatars.folder= ./avatars
 spring.jackson.serialization.FAIL_ON_EMPTY_BEANS=false
-spring.liquibase.change-log=classpath:liquibase/changelog-master.yml
