--- src/main/resources/liquibase/changelog-master.yml
@@ -1,3 +1,0 @@
-databaseChangeLog:
-  - include:
-     file: liquibase/scripts/student.sql
